+++ lucabook/lib/luca_book/dict.rb
@@ -2,12 +2,19 @@
 
 require 'luca_support/config'
 require 'luca_record/dict'
+require 'date'
+require 'pathname'
 
 module LucaBook
   class Dict < LucaRecord::Dict
+    def self.latest_balance
+      dict_dir = Pathname(LucaSupport::Config::Pjdir) / 'data' / 'balance'
+      # TODO: search latest balance dictionary
+      load_tsv_dict(dict_dir / 'start.tsv')
+    end
 
-    @filename = 'dict.tsv'
-
-    Data = load
+    def self.issue_date(obj)
+      Date.parse(obj.dig('_date', :label))
+    end
   end
 end